+++ server/src/main/java/org/elasticsearch/lucene/spatial/CartesianShapeDocValuesQuery.java
@@ -1,16 +1,16 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.spatial.index.mapper;
+package org.elasticsearch.lucene.spatial;
 
 import org.apache.lucene.document.ShapeField;
 import org.apache.lucene.geo.Component2D;
 import org.apache.lucene.geo.XYGeometry;
-import org.elasticsearch.lucene.spatial.CoordinateEncoder;
 
 /** Lucene geometry query for {@link BinaryShapeDocValuesField}. */
 public class CartesianShapeDocValuesQuery extends ShapeDocValuesQuery<XYGeometry> {
