+++ server/src/test/java/org/elasticsearch/lucene/spatial/LatLonShapeDocValuesQueryTests.java
@@ -1,18 +1,20 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.spatial.index.mapper;
+package org.elasticsearch.lucene.spatial;
 
 import org.apache.lucene.document.Document;
+import org.apache.lucene.document.LatLonShape;
 import org.apache.lucene.document.ShapeField;
-import org.apache.lucene.document.XYShape;
-import org.apache.lucene.geo.XYGeometry;
-import org.apache.lucene.geo.XYPolygon;
-import org.apache.lucene.geo.XYRectangle;
+import org.apache.lucene.geo.LatLonGeometry;
+import org.apache.lucene.geo.Point;
+import org.apache.lucene.geo.Polygon;
+import org.apache.lucene.geo.Rectangle;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
@@ -22,36 +24,35 @@ import org.apache.lucene.index.SerialMergeScheduler;
 import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.store.Directory;
+import org.apache.lucene.tests.geo.GeoTestUtil;
 import org.apache.lucene.tests.search.CheckHits;
 import org.apache.lucene.tests.search.QueryUtils;
-import org.elasticsearch.common.geo.LuceneGeometriesUtils;
+import org.elasticsearch.common.geo.Orientation;
 import org.elasticsearch.core.IOUtils;
-import org.elasticsearch.geo.ShapeTestUtils;
-import org.elasticsearch.geo.XShapeTestUtil;
+import org.elasticsearch.geo.GeometryTestUtils;
 import org.elasticsearch.geometry.Geometry;
-import org.elasticsearch.lucene.spatial.CartesianShapeIndexer;
-import org.elasticsearch.lucene.spatial.CoordinateEncoder;
+import org.elasticsearch.index.mapper.GeoShapeIndexer;
 import org.elasticsearch.test.ESTestCase;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.function.Function;
 
-public class CartesianShapeDocValuesQueryTests extends ESTestCase {
+public class LatLonShapeDocValuesQueryTests extends ESTestCase {
 
     private static final String FIELD_NAME = "field";
 
     public void testEqualsAndHashcode() {
-        XYPolygon polygon = LuceneGeometriesUtils.toXYPolygon(ShapeTestUtils.randomPolygon(false));
-        Query q1 = new CartesianShapeDocValuesQuery(FIELD_NAME, ShapeField.QueryRelation.INTERSECTS, polygon);
-        Query q2 = new CartesianShapeDocValuesQuery(FIELD_NAME, ShapeField.QueryRelation.INTERSECTS, polygon);
+        Polygon polygon = GeoTestUtil.nextPolygon();
+        Query q1 = new LatLonShapeDocValuesQuery(FIELD_NAME, ShapeField.QueryRelation.INTERSECTS, polygon);
+        Query q2 = new LatLonShapeDocValuesQuery(FIELD_NAME, ShapeField.QueryRelation.INTERSECTS, polygon);
         QueryUtils.checkEqual(q1, q2);
 
-        Query q3 = new CartesianShapeDocValuesQuery(FIELD_NAME + "x", ShapeField.QueryRelation.INTERSECTS, polygon);
+        Query q3 = new LatLonShapeDocValuesQuery(FIELD_NAME + "x", ShapeField.QueryRelation.INTERSECTS, polygon);
         QueryUtils.checkUnequal(q1, q3);
 
-        XYRectangle rectangle = XShapeTestUtil.nextBox();
-        Query q4 = new CartesianShapeDocValuesQuery(FIELD_NAME, ShapeField.QueryRelation.INTERSECTS, rectangle);
+        Rectangle rectangle = GeoTestUtil.nextBox();
+        Query q4 = new LatLonShapeDocValuesQuery(FIELD_NAME, ShapeField.QueryRelation.INTERSECTS, rectangle);
         QueryUtils.checkUnequal(q1, q4);
     }
 
@@ -65,14 +66,14 @@ public class CartesianShapeDocValuesQueryTests extends ESTestCase {
         // RandomIndexWriter is too slow here:
         IndexWriter w = new IndexWriter(dir, iwc);
         final int numDocs = randomIntBetween(10, 1000);
-        CartesianShapeIndexer indexer = new CartesianShapeIndexer(FIELD_NAME);
+        GeoShapeIndexer indexer = new GeoShapeIndexer(Orientation.CCW, FIELD_NAME);
         for (int id = 0; id < numDocs; id++) {
             Document doc = new Document();
             @SuppressWarnings("unchecked")
             Function<Boolean, Geometry> geometryFunc = ESTestCase.randomFrom(
-                ShapeTestUtils::randomLine,
-                ShapeTestUtils::randomPoint,
-                ShapeTestUtils::randomPolygon
+                GeometryTestUtils::randomLine,
+                GeometryTestUtils::randomPoint,
+                GeometryTestUtils::randomPolygon
             );
             Geometry geometry = geometryFunc.apply(false);
             List<IndexableField> fields = indexer.indexShape(geometry);
@@ -93,10 +94,10 @@ public class CartesianShapeDocValuesQueryTests extends ESTestCase {
 
         IndexSearcher s = newSearcher(r);
         for (int i = 0; i < 25; i++) {
-            XYGeometry[] geometries = randomLuceneQueryGeometries();
+            LatLonGeometry[] geometries = randomLuceneQueryGeometries();
             for (ShapeField.QueryRelation relation : ShapeField.QueryRelation.values()) {
-                Query indexQuery = XYShape.newGeometryQuery(FIELD_NAME, relation, geometries);
-                Query docValQuery = new CartesianShapeDocValuesQuery(FIELD_NAME, relation, geometries);
+                Query indexQuery = LatLonShape.newGeometryQuery(FIELD_NAME, relation, geometries);
+                Query docValQuery = new LatLonShapeDocValuesQuery(FIELD_NAME, relation, geometries);
                 assertQueries(s, indexQuery, docValQuery, numDocs);
             }
         }
@@ -113,10 +114,10 @@ public class CartesianShapeDocValuesQueryTests extends ESTestCase {
         // RandomIndexWriter is too slow here:
         IndexWriter w = new IndexWriter(dir, iwc);
         final int numDocs = randomIntBetween(10, 100);
-        CartesianShapeIndexer indexer = new CartesianShapeIndexer(FIELD_NAME);
+        GeoShapeIndexer indexer = new GeoShapeIndexer(Orientation.CCW, FIELD_NAME);
         for (int id = 0; id < numDocs; id++) {
             Document doc = new Document();
-            Geometry geometry = ShapeTestUtils.randomGeometryWithoutCircle(randomIntBetween(1, 5), false);
+            Geometry geometry = GeometryTestUtils.randomGeometryWithoutCircle(randomIntBetween(1, 5), false);
             List<IndexableField> fields = indexer.indexShape(geometry);
             for (IndexableField field : fields) {
                 doc.add(field);
@@ -135,10 +136,10 @@ public class CartesianShapeDocValuesQueryTests extends ESTestCase {
 
         IndexSearcher s = newSearcher(r);
         for (int i = 0; i < 25; i++) {
-            XYGeometry[] geometries = randomLuceneQueryGeometries();
+            LatLonGeometry[] geometries = randomLuceneQueryGeometries();
             for (ShapeField.QueryRelation relation : ShapeField.QueryRelation.values()) {
-                Query indexQuery = XYShape.newGeometryQuery(FIELD_NAME, relation, geometries);
-                Query docValQuery = new CartesianShapeDocValuesQuery(FIELD_NAME, relation, geometries);
+                Query indexQuery = LatLonShape.newGeometryQuery(FIELD_NAME, relation, geometries);
+                Query docValQuery = new LatLonShapeDocValuesQuery(FIELD_NAME, relation, geometries);
                 assertQueries(s, indexQuery, docValQuery, numDocs);
             }
         }
@@ -150,21 +151,21 @@ public class CartesianShapeDocValuesQueryTests extends ESTestCase {
         CheckHits.checkEqual(docValQuery, s.search(indexQuery, numDocs).scoreDocs, s.search(docValQuery, numDocs).scoreDocs);
     }
 
-    private XYGeometry[] randomLuceneQueryGeometries() {
+    private LatLonGeometry[] randomLuceneQueryGeometries() {
         int numGeom = randomIntBetween(1, 3);
-        XYGeometry[] geometries = new XYGeometry[numGeom];
+        LatLonGeometry[] geometries = new LatLonGeometry[numGeom];
         for (int i = 0; i < numGeom; i++) {
             geometries[i] = randomLuceneQueryGeometry();
         }
         return geometries;
     }
 
-    private XYGeometry randomLuceneQueryGeometry() {
+    private LatLonGeometry randomLuceneQueryGeometry() {
         return switch (randomInt(3)) {
-            case 0 -> LuceneGeometriesUtils.toXYPolygon(ShapeTestUtils.randomPolygon(false));
-            case 1 -> LuceneGeometriesUtils.toXYCircle(ShapeTestUtils.randomCircle(false));
-            case 2 -> LuceneGeometriesUtils.toXYPoint(ShapeTestUtils.randomPoint(false));
-            default -> XShapeTestUtil.nextBox();
+            case 0 -> GeoTestUtil.nextPolygon();
+            case 1 -> GeoTestUtil.nextCircle();
+            case 2 -> new Point(GeoTestUtil.nextLatitude(), GeoTestUtil.nextLongitude());
+            default -> GeoTestUtil.nextBox();
         };
     }
 }
